Moved filtering logic to util folder for easier maintenance per PR code review comment
